// server/src/interfaces/dataPoint.interface.ts
interface IDataPoint {
  iso_code: string;
  country: {
    alpha2: string;
    alpha3: string;
    countryCallingCode: string[];
    currencies: string[];
    ioc: string;
    languages: string[];
    name: string;
    status: string;
    // this is probably always a number but better to treat as a string
    iso_code: string;
  };
  year: string;
  coal_prod_change_pct: number;
  coal_prod_change_twh: number;
  gas_prod_change_pct: number;
  gas_prod_change_twh: number;
  oil_prod_change_pct: number;
  oil_prod_change_twh: number;
  energy_cons_change_pct: number;
  energy_cons_change_twh: number;
  biofuel_share_elec: number;
  biofuel_cons_change_pct: number;
  biofuel_share_energy: number;
  biofuel_cons_change_twh: number;
  biofuel_consumption: number;
  biofuel_elec_per_capita: number;
  biofuel_cons_per_capita: number;
  carbon_intensity_elec: number;
  coal_share_elec: number;
  coal_cons_change_pct: number;
  coal_share_energy: number;
  coal_cons_change_twh: number;
  coal_consumption: number;
  coal_elec_per_capita: number;
  coal_cons_per_capita: number;
  coal_production: number;
  coal_prod_per_capita: number;
  electricity_generation: number;
  biofuel_electricity: number;
  coal_electricity: number;
  fossil_electricity: number;
  gas_electricity: number;
  hydro_electricity: number;
  nuclear_electricity: number;
  oil_electricity: number;
  other_renewable_electricity: number;
  other_renewable_exc_biofuel_electricity: number;
  renewables_electricity: number;
  solar_electricity: number;
  wind_electricity: number;
  energy_per_gdp: number;
  energy_per_capita: number;
  fossil_cons_change_pct: number;
  fossil_share_energy: number;
  fossil_cons_change_twh: number;
  fossil_fuel_consumption: number;
  fossil_energy_per_capita: number;
  fossil_cons_per_capita: number;
  fossil_share_elec: number;
  gas_share_elec: number;
  gas_cons_change_pct: number;
  gas_share_energy: number;
  gas_cons_change_twh: number;
  gas_consumption: number;
  gas_elec_per_capita: number;
  gas_energy_per_capita: number;
  gas_production: number;
  gas_prod_per_capita: number;
  hydro_share_elec: number;
  hydro_cons_change_pct: number;
  hydro_share_energy: number;
  hydro_cons_change_twh: number;
  hydro_consumption: number;
  hydro_elec_per_capita: number;
  hydro_energy_per_capita: number;
  low_carbon_share_elec: number;
  low_carbon_electricity: number;
  low_carbon_elec_per_capita: number;
  low_carbon_cons_change_pct: number;
  low_carbon_share_energy: number;
  low_carbon_cons_change_twh: number;
  low_carbon_consumption: number;
  low_carbon_energy_per_capita: number;
  nuclear_share_elec: number;
  nuclear_cons_change_pct: number;
  nuclear_share_energy: number;
  nuclear_cons_change_twh: number;
  nuclear_consumption: number;
  nuclear_elec_per_capita: number;
  nuclear_energy_per_capita: number;
  oil_share_elec: number;
  oil_cons_change_pct: number;
  oil_share_energy: number;
  oil_cons_change_twh: number;
  oil_consumption: number;
  oil_elec_per_capita: number;
  oil_energy_per_capita: number;
  oil_production: number;
  oil_prod_per_capita: number;
  other_renewables_elec_per_capita: number;
  other_renewables_elec_per_capita_exc_biofuel: number;
  other_renewables_share_elec: number;
  other_renewables_share_elec_exc_biofuel: number;
  other_renewables_cons_change_pct: number;
  other_renewables_share_energy: number;
  other_renewables_cons_change_twh: number;
  other_renewable_consumption: number;
  other_renewables_energy_per_capita: number;
  per_capita_electricity: number;
  population: number;
  primary_energy_consumption: number;
  renewables_elec_per_capita: number;
  renewables_share_elec: number;
  renewables_cons_change_pct: number;
  renewables_share_energy: number;
  renewables_cons_change_twh: number;
  renewables_consumption: number;
  renewables_energy_per_capita: number;
  solar_share_elec: number;
  solar_cons_change_pct: number;
  solar_share_energy: number;
  solar_cons_change_twh: number;
  solar_consumption: number;
  solar_elec_per_capita: number;
  solar_energy_per_capita: number;
  gdp: number;
  wind_share_elec: number;
  wind_cons_change_pct: number;
  wind_share_energy: number;
  wind_cons_change_twh: number;
  wind_consumption: number;
  wind_elec_per_capita: number;
  wind_energy_per_capita: number;
}